+++ scripts/dev-bundle-tool-package.js
@@ -14,9 +14,11 @@ var packageJson = {
     pacote: "https://github.com/meteor/pacote/tarball/c5043daa1b768594e01d76275e3854fc19f038f9",
     "node-gyp": "3.7.0",
     "node-pre-gyp": "0.10.3",
-    "meteor-babel": "7.4.4",
+    "meteor-babel": "7.4.5",
+    // Keep the versions of these packages consistent with the versions
+    // found in dev-bundle-server-package.js.
     "meteor-promise": "0.8.7",
-    reify: "0.17.3",
+    reify: "0.18.1",
     fibers: "3.1.1",
     // So that Babel can emit require("@babel/runtime/helpers/...") calls.
     "@babel/runtime": "7.4.3",
